pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "fiscal-de-filmes"
version = "0.1.0"
description = ""
authors = [
    {name = "Mariana Barreto",email = "marianaportobarreto@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests (>=2.31.0)",
    "beautifulsoup4 (>=4.12.0)",
    "lxml (>=4.9.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)",
    "psycopg2-binary (>=2.9.9)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "fastapi (>=0.111.0)",
    "uvicorn (>=0.29.0)",
    "httpx (>=0.27.0)",
    "pydantic (>=2.7.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
